extracting board logic
